puff: implement PP5000 port power control
The EC watches current limiting on the USB and HDMI ports to keep system
load on the 5V rail within the limits of the regulator. The EC's goal is
to ensure that any single future state change (connecting or
disconnecting a port) doesn't push the 5V rail overcurrent.

Verification is done with a Python script that explores the state space
and checks that the EC never permits overload, by implementing the same
control logic in apply_ec_control() and update_port_limits().

BUG=b:143190102
TEST=make buildall
BRANCH=none

Change-Id: If06e319c8d38bd11b29a7e69499d40357176a97e
Signed-off-by: Peter Marheine <pmarheine@chromium.org>