puff: implement USB-A port power control
It's safe to rely solely on interrupts for twiddling USB_A_LOW_PWR_OD
because the USB and HDMI power rails are EC-controlled and default off,
so HDMI_CONN*_OC_ODL can only be asserted after we're ready to service
the interrupts.

USB_A_LOW_PWR_OD forces low-power when high and it has an external
pull-up, so we change the output type to ODR_LOW to default to full
power.

BUG=b:143190102
TEST=make buildall
BRANCH=none

Change-Id: If06e319c8d38bd11b29a7e69499d40357176a97e
Signed-off-by: Peter Marheine <pmarheine@chromium.org>