Update patch set 6
Patch Set 6:

Pardon the nit, as this suggestion is merely a future TODO:

pe[port].sender_response_timer == TIMER_DISABLED

From what I understand, this check is used to ensure we're not waiting for a response for a message before replying. (SRC_CAP=> REQUEST, etc)

Does this code check if it is the [first message in an Uninterruptable AMS], or instead an [Interruptable AMS]? The spec says certain AMS should be "abandoned" if a higher-priority non-interruptible AMS comes through.

> 6.4.4.5 VDM Message Timing and Normal PD Messages
> Any Command Process or other VDM sequence May be interrupted by any other USB PD Message. The Vendor or Standards defined state operation Shall comprehend this and continue to operate as expected when processing any other USB PD Messages.

8.3.2 Atomic Message Sequence Diagrams
Section 8.3.2.1.3 gives details of which of these AMS’s are interruptible or non-interruptible.

At some point in the future, we may want to allow for "interruptability" per the table in the spec.

Patch-set: 6