stm32mon: Add connect retries parameter
Currently, stm32mon will try to get the attention of the
stm32 bootloader for forever.
Retrying forever is odd behavior, in general.
Furthermore, this behavior does not allow wrapper scripts to
observe the failure and try other retry mechanisms, like
resetting the chip.

TL;DR
This CL sets a default 40 retries for the bootloader connect/sync
and adds the "--retries <num-retries>" and "-R <num-retries>" argument
for adjustments.
Note, this is simply the retry limit for connecting/syncing-with the
bootlodaer.
You can specify "-R -1" to try forever.

BRANCH=none
BUG=b:143374692
TEST=cros_workon --board=nocturne start chromeos-base/ec-devutils
     emerge-nocturne chromeos-base/ec-devutils
     cros deploy --root=/usr/local dut1 chromeos-base/ec-devutils
     # Bind spidev for fpmcu
     # Enable hardware wp
     time stm32mon -U -u -p -s /dev/spidev32765.0 -e -w \
         /opt/google/biod/fw/nocturne_fp_v2.2.110-b936c0a3c.bin
     # Check that it tries forever??
     time stm32mon -C 10 -U -u -p -s /dev/spidev32765.0 -e -w \
         /opt/google/biod/fw/nocturne_fp_v2.2.110-b936c0a3c.bin
     # Check that is only makes 11 attempts

Change-Id: I925466fb892b0d17a7ff8b354e3cee302d167d3c
Signed-off-by: Craig Hesling <hesling@chromium.org>