Update patch set 3
Patch Set 3: Code-Review+2

The code looks fine insofar as I understand what it should be doing.
This is a complex sequence, and I think it would be great if there was a go/comet-lake-power
link to doc that described just what is going on, and the meaning of the signals.
In a general sense I also worry about GPIO naming - there doesn't seem to be a consistent
naming of power signals in common code, and this leads to the
EC schematic name being aliased to some name used in common code.
Aseda should also review.

Patch-set: 3
Reviewer: Gerrit User 1281040 <1281040@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2