kohaku: set battery LED green when not charging due to overtemp
When charging stops due to battery overtemperature, the battery LED
should be green regardless of the state of charge of the battery.
The logic checks the battery temperature against the maximum charging
temperature, and if it is too high, it sets the charging state
machine to ST_IDLE. However, if the battery temperature is above the
maximum discharging temperature, then the charging state machine
will shut down the system completely.

BUG=b:140596424
BRANCH=None
TEST=`make -j BOARD=kohaku && ./util/flash_ec --board=kohaku`
Ensure battery is not fully charged
LED is blue
Plug in AC adapter
LED is still blue
Close lid
LED turns red
Access EC console
> batttempfake 3300
This overrides the normal temperature read and forces the value to
330.0 K, which is 56.9 C. The Kohaku battery has a maximum charge
temperature of 55 C, and a maximum discharge temperature of 60 C,
so this temperature is hot enough to stop charging, but not so hot
that the system will shut down.
LED turns green
> batttempfake -1
LED turns red

Change-Id: I3b76eab942ca3ef3871f0909395e91634db5640e
Signed-off-by: Paul Fagerburg <pfagerburg@chromium.org>