driver: fix lsm6dsm polling mode support problem
Need to use lsm6dsm polling mode on tglrvp, but lsm6dsm doesn't
correctly handle the case that the CONFIG_ACCEL_INTERRUPTS and
the CONFIG_ACCEL_FIFO are not defined which is the case for
polling mode. For example, fifo_disable is defined when the
CONFIG_ACCEL_INTERRUPTS is set but used under the if clause of
IS_ENABLED(CONFIG_ACCEL_FIFO) which is different with the
ifdef CONFIG_ACCEL_FIFO and the compiler will compile these code
although CONFIG_ACCEL_FIFO is not defined, then build fail. This
patch is my solution for this problem but any other better solution
is also ok for me.

BUG=none
BRANCH=none
TEST=tested on tgl rvp

Signed-off-by: Leifu Zhao <leifu.zhao@intel.com>
Change-Id: I5b338d81061f25fd1c8209b4555f63ea4d8b2dbc