Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > The common charging state machine or the fuel gauge itself doesn't stop charging once the temp limits are reached?
> 
> Nope. charge_state_v2.c has battery_too_hot(int batt_temp_c), which will check if the battery temperature exceeded the limits for discharging, and if so, shut down the whole system, but the charge state machine does not pay attention to the charging temperature range.
> 
> Also, battery_too_hot will return a true value if the battery temperature is *below* the minimum discharging temperature :-/

This change was motivated by desired LED behavior and the fix is to have the charging temperature limits be properly used. So that makes sense. My question is whether this should be available to or useful in general? Could this be moved the charge_state_v2 and guarded with a config option if we're concerned about changing charging behavior by default?

Patch-set: 2
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>