flash_fp_mcu: Reset line open/drain handling
This change ensures that we are handling the FPMCU reset line
in an open/drain manner, for designs that have the AP GPIO
directly connected to the FPMCU's nRST pin.

This is important because STM32H743 can drive this reset line in
certain situation, I believe POR.

I have not verified if the AP is using a purely open/drain driver
config for this pin, but this fix ensure correctness in either situation.

BRANCH=nocturne,hatch
BUG=b:143374692
TEST=# Ensure PS crrev.com/c/1921705 is applied.


Change-Id: I9d9ac81953b030ffa32dc7a74f019f319be4c764
Signed-off-by: Craig Hesling <hesling@chromium.org>