Update patch set 32
Patch Set 32:

> The I2C access pretty early will not be a problem 
>in the chip of ST. But, if you use other EC chip, 

This is good to know. Let's open a bug for this. For now let's be happy that it's working well in STM32. We'll we'll cross that bridge when we come to it in terms of this driver on other ECs.

Patch-set: 32
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0