TCPMv1: Improve jitter based collision avoidance
For the holdoff timer, we were using the system timestamp as a source of
adding jitter.  However, it seems that there are certain cases that can
cause a periodic collision with the jitter only varying by ~1-2ms.  This
leads to many repeated collisions until the jitter slides just out of
the collision window.  This commit changes the jitter calculation to use
more of the bits that are changing to reduce the number of collisions.

BUG=b:144676183, chromium:925618
BRANCH=hatch,atlas,nocturne,grunt,octopus,rammus,kukui
TEST=Flash kohaku, plug some peripherals known for collisions, verify
that if a collision is encountered, it's only encountered once (no
cycles).
TEST=Add debugging prints for the jitter, verify that the jitter is
varied and appears non-deterministic.

Change-Id: I6c27880551dd35b78993f7130d1ce4eb81aa10ef
Signed-off-by: Aseda Aboagye <aaboagye@google.com>