Update patch set 1
Patch Set 1:

In light of Security Week, James dropped by the lab and commented this:
"In this context, a Multiply-with-Carry PRNG has much better periodicity."

He then opened a text editor, typed this pseudocode, and mic-dropped:
"u32 state

u32 rand_next(u32 * state)
	*state = *state * prime1 + prime2
	return state"

I defer to embedded specialists such as he when it comes to this, so posting for posterity.
Aseda's solution works well enough for our limited-processor ECs.

Also, James asked if we have a "TRNG" to read from.
I defer to him on that, as AFAIK this code is EC-chipset agnostic.

Patch-set: 1
CC: Gerrit User 1314041 <1314041@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1269501 <1269501@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>