Update patch set 1
Patch Set 1: Code-Review+1

I like the code change but your commit comment seems over specific and/or misleading. The 480 was previously determined and not a change for this CL.  Not sure it is required to be commented upon here. It looks more like the problem this CL is solving is it uses 32 bits for

Patch-set: 1
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1