cr50: switch rand_bytes() to use HMAC SHA256 DRBG
An “Approved” RNG listed in FIPS 140-2 Annex C must be used for the
generation of random data or cryptographic keys used by an approved
security function. Detailed information and guidance on Key Generation
in FIPS 140-2 can be found in NIST SP 800-133 and FIPS 140-2 IG 7.8.

Currently rand() and rand_bytes() use raw TRNG data, which is not
compliant with FIPS 140-2. With this change following features are
implemented:
  1. introduced get_entropy() and get_entropy32() functions which
     gather enough entropy from TRNG to produce full entropy via
     condensing entropy with approved external conditioning SHA256
  2. added global HMAC DRBG which is initialized on first use with
     440 bits of entropy as required by NIST SP 800 90A rev 1
  3. implemented proper reseeding of global DRBG with 256 bits of
     full entropy.
  4. fips_rand() and fips_rand_bytes() are implemented using DRBG.

BUG=b:138577416
BRANCH=cr50
TEST=nist_entropy.sh && tpmtest.py -t1 && tpmtest.py -t2 to make sure
that introduced changes really produce random numbers.

Change-Id: Ib7b80871bde76f091e8f8e48e473f73bd77bcdad
Signed-off-by: Vadim Sukhomlinov <sukhomlinov@google.com>