Update patch set 25
Patch Set 25:

(5 comments)

The only work that I see that needs to be done is for the case when another task, like the host command task, talks to the TCPC while the PD task and state machine is in low power mode. In this scenario the PD task needs to be able to wake back up and put the TCPC back to low power mode after the host command task is done communicating with the TCPC (the low power mode wait timer)

you can test this by using ectool on the AP console to interact with the TCPC like ectool pdchipinfo or something like that.

Patch-set: 25
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>