Nami: USB-PD: check dual-role status in S0ix/S5 and add delay time for dual-role device
Charging via hub with the ANX7447 TCPC fails when the system is in
s0ix/s5. Both the hub and the ANX7447 are auto-toggling, so source and
sink roles depend on the timing of the connection. Charging will fail
when the hub expects the ANX7447 to be a source, but the drp_state in
S0ix/s5 is TOGGLE_OFF/FORCE_SINK.

Ideally we wouldn't use auto-toggle when drp_state is
TOGGLE_OFF/FORCE_SINK, but with the ANX7447 TCPC, auto-toggle can't be
prevented in low power mode.

To fix this, try being a sink in case the connected device is dual-role.
100 ms is enough time for a dual-role partner to switch from sink to
source. If the connected device is sink-only, then we will attempt
SNK_DISCONNECTED twice (due to debounce time), then return to low power
mode (and stay there). After 200 ms, reset ready for a new connection.

Move the next_state selection out into a function since things were
getting very narrow inside PD_STATE_DRP_AUTO_TOGGLE.

BUG=b:143917248
BRANCH=nami
TEST=verify Sona G7  s5/s0ix plug-in dual-role device can be charging

Change-Id: Ic7e4a74724af4d20a34c299cf7e780fb15cf4c25
Signed-off-by: Leo-Tsai <leocx_tsai@compal.corp-partner.google.com>