Update patch set 2
Patch Set 2:

Does it prevent the OS from detecting the devices on the USBC port?

I have been observing a behavior where the device on the USBC port is not detected automatically after bootup. The behavior is as follows:

1) I have connected a USBC to ethernet dongle on the USBC port
2) Ran a script which repeatedly performs a warm reset i.e. reboot AP only
3) After certain iterations of reboot, the USB dongle is not getting detected. lsusb on the OS side does not show anything.
4) I manually unplug and replug the dongle to get it working again.

This issue has been observed on both the USBC ports (MB and DB). Can this CL explain that behavior or is that totally different?

Patch-set: 2
Reviewer: Gerrit User 1302633 <1302633@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>