Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> The rev0 board (Jacuzzi ref. board, not any OEM boards) doesn't include these part of pp1800_s5 control circuit, since we caught and implement it after Jacuzzi proto. And we had to repurpose PA14 due to insufficient pins on EC

The problem is not the repurposing, they happen from time to time. But it's a delicate matter which pin you do this on, in case you run old firmware on new boards or new firmware on old boards (like in this case).

The problem now is that the current firmware is assuming it can freely output on PA14, a pin which on rev 0 is an input (and the lid accelerometer might be outputting on that net too). So there's a chance we have a short now.

Patch-set: 9