charger/rt946x: do BC1.2 recognition when required
BC1.2 and USB enumeration may have a race condition on
operating D+/D- pins. This CL make BC1.2 detection runs
only when required. That is, only if the CC pins exists RP (sink),
and the port is not PD capable, then the source could be a
BC1.2 device.

Also, drops notifying AC_CHANGE, since it's not accurate while
we only enable BC1.2 detection on the state.

TEST=ensure BC1.2 detection only runs on a BC1.2 charger plug.
BUG=b:141005922
BRANCH=kukui

Change-Id: I0aac4221858b316165302494933bf0f54809dcf1
Signed-off-by: Yilun Lin <yllin@chromium.org>
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>