trembyle: POWER_CTRL auto discharge disconnect init
The POWER_CTRL Auto Discharge Disconnect bit is supposed
to default to 0.  There is a situation when the AP is not
running that this bit is set and will not allow "AC on"
to happen.  Clearing this bit at initialization to be the
default state of this bit fixes this issue.

BUG=b:144126745
BRANCH=none
TEST=Charger attach/pull with AP not running

Change-Id: I625efbba80f190322e3e92de6318b710b3ce7ade
Signed-off-by: Denis Brockus <dbrockus@chromium.org>