tcpci: restructure fault fetch and clearing
Changed the code to not be as indented by using functions
instead of nesting deeper and deeper in conditions.  The
intent would be to add tcpci_handle_fault(port, fault) if
we ever decide to generically handle faults.  Then this
would be included in the if (get==SUCCESS && clear=SUCCESS)
condition.

BUG=b:144126745
BRANCH=none
TEST=if any console FAULT output verify properly handled

Change-Id: Ie406302c4a824e7d14ca0832cf7bfca7432680f6
Signed-off-by: Denis Brockus <dbrockus@chromium.org>