usb_pd_policy: Make a lot of objects common
There is a board specific usb_pd_policy.c file that contains a lot of
code for handling DisplayPort Alternate mode, Google Firmware Update
Alternate mode, as well as some PD policy functions such as deciding to
Accept or Reject a data role swap or a power role swap.  Several boards
simply copy/paste this code from project to project as a lot of this
functionality is not actually board specific.

This commit tries to refactor this by pulling the functions that are not
mainly board specific into common code.  The functions are made
overridable such that boards that truly do require a different
implementation may do so.

In order to facilitate this, a couple CONFIG_* options were introduced:

- CONFIG_USB_PD_DP_HPD_GPIO
/* HPD is sent to the GPU from the EC via a GPIO */

- CONFIG_USB_PD_CUSTOM_VDO
/*
 * Define this if a board needs custom SNK and/or SRC PDOs.
 *
 * The default SRC PDO is a fixed 5V/1.5A with PDO_FIXED_FLAGS indicating
 * Dual-Role power, USB Communication Capable, and Dual-Role data.
 *
 * The default SNK PDOs are:
 * - Fixed 5V/500mA with the same PDO_FIXED_FLAGS
 * - Variable (non-battery) min 4.75V, max PD_MAX_VOLTAGE_MV,
 *   operational current PD_MAX_CURRENT_MA,
 * - Battery min 4.75V, max PD_MAX_VOLTAGE_MV, operational power
 *   PD_OPERATING_POWER_MW
 */

BUG=chromium:1021724
BRANCH=<as many as we can that are still supported>
TEST=`make -j buildall`
TEST=Flash a kohaku, verify that DP Alt Mode still works with a variety
of DP peripherals
TEST=Repeat above with a nocturne
TEST=Repeat above with an atlas

Change-Id: I18fd7e22dc77fe1dc6c21c38cd7f1bc53cae86cb
Signed-off-by: Aseda Aboagye <aaboagye@google.com>