npcx: Disable ITIM after watchdog_stop_and_unlock()
watchdog_stop_and_unlock() depends upon the ITIM32 module to ensure
that it has been at least 3 watchdog ticks since the last time the
watchdog has been touched.  If it has been > 100ms since then, there
will be no problem, but if it has been less than 100ms, then because
the ITIM32 module was disabled, then the system will get stuck in the
while (time_since32() ...) loop.  This will eventually cause the
watchdog to kick in and reboot the EC, which will also cause the
AP to reboot.

BUG=b:145371494, b:140207603, b:462416314
BRANCH=firmware-hatch-12672.B
TEST=sanity testing (alt+volup+h still works) and
buildall (this bug was nearly impossible to reproduce)

Change-Id: I5e32f4940f03fee90b3aa809b185c13ed66ce7f1
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1946773
Reviewed-by: Andrew McRae <amcrae@chromium.org>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>