gsctool: add 'force_ro' option
By design gsctool does not update the inactive RO unless the running
RO version is lower than the version in the Cr50 image being
transferred.

With transition to the new RO version, keys and rollback map state,
there is a need to enforce that both ROs on the device are updated.

This patch adds a command line option which allows to update the
inactive RO to the contents of the Cr50 image.

BRANCH=none
BUG=b:136284186
TEST=running 'gsctool -q cr50.bin' updates the inactive RO section.

Change-Id: I79c21fc2cd41c40070352d8fbce0b71dd3ae2b1f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1949547
Reviewed-by: Namyoon Woo <namyoon@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
(cherry picked from commit 53f59f6b361e8cf0a8d1af70306146db90ffae23)