build_assert: use _Static_assert
We have compiler support for _static_assert, so we should use it. Also
allow for an option message in BUILD_ASSERT that will display as a part
of the failure.

Here is an example of a compile failure with the optional message

...
  CC      RW/chip/npcx/gpio.o
  CC      RW/chip/npcx/header.o
In file included from include/common.h:13,
                 from include/adc.h:11,
                 from board/phaser/board.c:8:
include/compile_time_macros.h:13:2: error: static assertion failed:
"board/phaser/board.c:74: Size mismatch"
  _Static_assert(cond, file ":" #line ": " msg)
...

BRANCH=none
BUG=none
TEST=verified by manually update BUILD_ASSERTS and viewing output.

Change-Id: I6421084868ea81be8b3fa6ad0613730f4e4e149b
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1950820
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>