hammer: Change the KEY_RO fmap size
Currently, we set the KEY_RO fmap size to 1024 by default.
However, the size can't match with total_size in the struct vb21_struct_common.
If we want to match the total_size, we need to change CONFIG_RO_PUBKEY_SIZE
to 832 = 816 + 16 where 816 is the public key size and 16 is the header size

BUG=chromium:1030917
TEST=the firmware can be updated to masterball

Change-Id: I4903b4db099c785e7b3446686c7d6a50b4ba2500