Update patch set 3
Patch Set 3: Code-Review+2

Please mark 'verified' patches which are ready for review.

As a result of this patch build/cr50/cryptoc/libcryptoc.a is not being built even though it is not required (cr50 uses LTO and links all objects directly), but this is a small price to pay for makefiles unification.

Patch-set: 3
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2