docs/fingerprint: Fix diagrams in fingerprint design doc
The doc links are not accessible publicly, so add SVG versions of the
images using Drive's export functionality (i.e., append "export/svg" to
the end of the link). The original doc links are kept as comments so we
know where they came from if we need to make changes.

BRANCH=none
BUG=none
TEST=view in gitiles

Change-Id: Ic6cf514032cfa9205c7c763716bf489597f2e08d
Signed-off-by: Tom Hughes <tomhughes@chromium.org>