usb_pd: Rename "Externally powered" to "Unconstrained Power"
This was changed in PD 2.0 years ago (via ECN authored by our own David
Schneider), but our codebase still refers to BIT 27 of the Fixed PDO as
"Externally powered" instead of "Unconstrained Power".

This will search and replace all instances of Externally powered" when it
refers to BIT 27, as well as function names, other internal representations
of that property, strings, and comments.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chromium:1030990
TEST=Codebase builds clean. No functional change, except for the property shows
up in ectool as "Unconstrained power" now.

Change-Id: I5ececa03f29eb31057be3d0ad5311117093bc6da