Trogdor: On startup, turn off the system without check
The original check in the init hook (power_chipset_init) may cause
watchdog reset, as the init hook is earlier than the task execution
that we should not perform any long action, like waiting for a signal
in our case.

We should simply turn off the switchcap without any check.

In most of the common cases (except flashing EC/AP), the switchcap
should be off and the check is unnecessary.

BRANCH=None
BUG=b:145843686
TEST=Performed flashing EC and AP, EC watchdog reset not seen.

Change-Id: I36873e773800def7e3dfceaec28c294dee9a09c7
Signed-off-by: Wai-Hong Tam <waihong@google.com>