g: eraseflashinfo: use old process to info1
The release branches use a different process than tot to read info1.
crrev.com/c/1928311 changed eraseflashinfo to use part of the tot method
to read info1, but not all variables are defined and the info1_layout
struct doesn't actually exist in the release branches, so we can't use
the tot method.

This change goes back to the old method of reading info1 instead of
switching to the tot method to limit the scope of the change for the mp
release. If we want to revisit unifying the tot and release branch
eraseflashinfo commands, we can later.

This change doesn't have any impact on images built without CR50_DEV=1.

BUG=b:145937858
BRANCH=cr50,cr50-mp
TEST=make -j BOARD=cr50 CR50_DEV=1

Change-Id: I382923307d94f9fe260cae4329735ff0bc3736fa
Signed-off-by: Mary Ruthven <mruthven@chromium.org>