Charge state v2: Only shutdown on cold discharge
Currently, the discharge minimum temperature is always checked when
determining if the system should shut down, even if the system is not
discharging.  This change allows the system to run below the discharging
minimum temperature as long as AC is present.

BRANCH=octopus
BUG=b:145494158
TEST=loaded onto octopus unit, faked low battery temperature and
verified the system did shut down when discharging and did not with AC

Change-Id: I4f6549b04dc1e7b55e410d3ae4f67dc6126f9c7c
Signed-off-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1958853
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>