puff: support switching between USB-C and barrel jack power
We let the charge controller decide which port to use, watching the
presence signal to tell the charge control what is connected. We only
allow switching the input when the system is off, since switching may
cause us to brown out.

This is mostly copied from Fizz; I'm unsure if we need to do anything
with the TCPPC when switching (Fizz has a GPIO that physically connects
or disconnects the type-C port).

BUG=b:143975429
TEST=booted on hardware, verified chgsup output with barrel jack
BRANCH=None

Change-Id: I6d2f8365c03aa44f272195708dd534fdc35cd2f6
Signed-off-by: Peter Marheine <pmarheine@chromium.org>