glados/chell: Stop building these boards
GLaDOS and chell are out of space, so stop building these boards now.
Deleting the boards is problematic since the ec-utils ebuild still
expects the board directory to exist, but only builds the utils-host
target. In order to prevent that, the 'cros_ec' USE flag needs to be
removed from their bsp as well.

BUG=None
BRANCH=None
TEST=`make -j buildall`
TEST=`cros_workon --board chell start ec-utils && emerge-chell ec-utils`

Cq-Depend: chrome-internal:2296525, chrome-internal:2295865
Change-Id: I249bb2b3957d2f237a63207c91f0bd2e973d5dca
Signed-off-by: Aseda Aboagye <aaboagye@google.com>