nct38xx: Set pull on both CC lines on disconnect
tcpci_nct38xx_set_cc() only sets the pull resistor on one CC line according
to polarity. This is correct when attached, but on disconnect we need to
set the pull resistor on both CC lines, since polarity is no longer known
(unless DRP toggle is enabled, since that will take care of setting both
CC lines to do the toggling).

This workaround for chromium:951681 can be removed once that bug is fixed
in the TCPCI commmon code.

BUG=b:146003980, chromium:951681
BRANCH=none
TEST=Charging works with both plug orientations with AP off

Change-Id: I1406263011a4c8d595be0d7093f2ab220690de3f
Signed-off-by: Edward Hill <ecgh@chromium.org>