Update patch set 1
Patch Set 1:

If this were a long-term change that affected multiple projects, it might be more appropriate to support all of the current limit options and control the chosen one with a config option. I can still do that, but I thought I would opt for simplicity, since Volteer is the only active project using this part, and the change is a temporary workaround.

Patch-set: 1