tcpci: push localized tcpci changes back into tcpci
Some of the NCT38XX driver code was more generic TCPCI
than specific to the chip.  So moved a number of the
functions over to be gernerically handled.

In doing this I propagated the idea of tri-state
polarity.  It is now either NORMAL, FLIPPED, or NONE
for unattached.  This is needed for the generic handling
of correctly setting CC.

BUG=b:146003980, chromium:951681
BRANCH=none
TEST=Charging works with both plug orientations with AP on
TEST=Device works with both plug orientations with AP on
TEST=Charging works with both plug orientations with AP off

Change-Id: Ie4b5cc998902a346a4f4a2c1480204b3a81017dd
Signed-off-by: Denis Brockus <dbrockus@chromium.org>