tcpci: change CC functions to work generically
I moved some of the code that was NCT38xx specific over
to TCPCI.  I also pulled in a few of the changes Ed made
in order to not have to create an extra CL of removing
his to put them in the generic location.  The code is
basically the same as it was in his.  I will investigate
a little deeper to make sure this is how we want to solve
this problem but in the mean time this will get us
back to functional on the zork platform.

BUG=b:146003980, chromium:951681
BRANCH=none
TEST=Charging works with both plug orientations with AP on
TEST=Device works with both plug orientations with AP on
TEST=Charging works with both plug orientations with AP off

Change-Id: Ie4b5cc998902a346a4f4a2c1480204b3a81017dd
Signed-off-by: Denis Brockus <dbrockus@chromium.org>