Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > If this were a long-term change that affected multiple projects, it might be more appropriate to support all of the current limit options and control the chosen one with a config option. I can still do that, but I thought I would opt for simplicity, since Volteer is the only active project using this part, and the change is a temporary workaround.
> 
> Do you think it would be better to just include this in the TCPC init of baseboard.c? Since we know we're removing it and we don't have to touch the common drivers at all.

It might be. I put it in the driver, because I was already looking in that part of the code for the field and register constants. It also has the advantage of obviously happening at the right time. I've now spent a few minutes trying to figure out when TCPC init happens relative to PPC init in each stack without getting to the bottom of it. I'm sure I'll be able to, but it's not as clear.

Patch-set: 1