Update patch set 1
Patch Set 1:

> Patch Set 1:
> It also has the advantage of obviously happening at the right time. I've now spent a few minutes trying to figure out when TCPC init happens relative to PPC init in each stack without getting to the bottom of it. I'm sure I'll be able to, but it's not as clear.

baseboard_tcpc_init is a hook called during EC initialization, whereas the PPC init function in the driver is called early in pd_task in the new stack and fairly early in pd_task in the old stack. (AFAIK, tasks start after hooks. I find it odd that we are enabling interrupts from devices that we haven't configured yet, but maybe I'm missing something.) The accesses to this register in the driver are all read-modify-write, so it should be safe to write to it earlier than driver init.

Patch-set: 1