Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What was the fix for the servo branch to make this work?

I've realized that my testing was invalid, I thought that the Servo branch was closer to master. This is safe to merge on master, but it is not safe to merge on the servo branch.

I'd like to have a discussion about synchronizing the servo branch in with master and people's concerns over it as there's a significant divergence in terms of total lines of code, but I haven't run into any issues testing builds off master since at least July so I am uncertain as to if the initial concern that caused the divergence has been resolved or if the problem is in a region that I don't heavily exercise.

Patch-set: 2
Reviewer: Gerrit User 1347380 <1347380@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0