npcx: build in one i2c xfer retry
The calling code was having to retry i2c requests when
there was an SMB Bus issue.  In most cases these were
cleaned up but it was left for the caller to retry the
transfer.  I moved the retry into the npcx code instead
to minimize duplicated code through out all callers.

BUG=none
BRANCH=none
TEST=check I2C requests to NPCX work properly

Change-Id: I0f2f3ca43a070c497e73ca40c824668e1be07fc8
Signed-off-by: Denis Brockus <dbrockus@chromium.org>