add BOARD_EC_CR50_COMM_SUPPORT to board properties
This CL add a board property indicating EC-CR50 communication
support. The target boards are Volteer,Dedede,Puff, and Zork.
It shall be detected if the H1 strap configuration value is
either 0x0E or 0xE0.

BUG=b:146567516, chromium:1027660
BRANCH=cr50
TEST=Flashed AP firmware through CCD on Grunt, Octopus, Scarlet
and Atlas.
This is the captured console log:
    --- UART initialized after reboot ---
    ...
    strap pin readings: a1:2 a9:2 a6:0 a12:0
    [0.005886 Valid strap: 0xa properties: 0x41]
    > brdprop
    properties = 0x1141

Flashed AP firmware on a reworked board with 1M ohm on DIOA1 and
5k ohm on DIOA9.
This is the captured console log:
    --- UART initialized after reboot ---
    ...
    strap pin readings: a1:2 a9:3 a6:0 a12:0
    [0.005886 Valid strap: 0xe properties: 0x200041]
    > brdprop
    properties = 0x201141

Change-Id: If60765190a385a0e728177911b1ec738c6a00d99
Signed-off-by: Namyoon Woo <namyoon@chromium.org>