stm32h7x3: Add backup regs to memory regions
The primary use of this is for clearing all memory regions
on boot. This is implemented in crrev.com/c/1949049.

It is okay to have this region in the linker script,
since you could technically assign variables to this
backup reg region without an issue. If this were to
happen, you would probably want to disable the existing
backup register mechanism.
Do keep in mind that this memory region must be enabled
on boot (like other regions).

BRANCH=nocturne,hatch
BUG=b:143471027
TEST=Tested with crrev.com/c/1949049

Change-Id: Iaa7520cdfc06d77a57bf7c7b316e6779c98aeca6
Signed-off-by: Craig Hesling <hesling@chromium.org>