stm32f4: Add backup regs to memory regions
The primary use of this is for clearing all memory regions
on boot. This is implemented in crrev.com/c/1949049.

It is okay to have this region in the linker script,
since you could technically assign variables to this
backup reg region without an issue. If this were to
happen, you would probably want to disable the existing
backup register usage.
Do keep in mind that this memory region must be enabled
on boot (like other regions).

BRANCH=hatch
BUG=b:143471027
TEST=none

Change-Id: I87ae4f65fa84f4416aa7fe58779a3f3e29b9bb7e
Signed-off-by: Craig Hesling <hesling@chromium.org>