nucleo-f412zg: Initial board offering
This brings initial support for the Nucleo-F412ZG
development board.

This configuration is similar to bloonchipper, but without
fingerprint sensor support, without rollback secret
support, and with an adjusted console USART selection.

The included README.md shows some quick steps to
get a Nucleo-F412ZG up and running EC.

Note that the RSA key included is only for test builds.
It should not be used for any production builds.

BRANCH=none
BUG=none
TEST=make BOARD=nucleo-f412zg
     make BOARD=nucleo-f412zg flash
     minicom -D/dev/ttyACM0

Change-Id: Ic1150152adbef9ed3f81aafb007be747b2c436af